# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Cultour Project",
        "title": "Cultour API",
        "contact": {
            "name": "Cultour Development Team"
        },
        "version": "1.0"
    },
    "host": "localhost:8181",
    "paths": {
        "/ask": {
            "post": {
                "description": "Generate text using Gemini AI",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "Generate AI text response",
                "parameters": [
                    {
                        "description": "AI Text Generation Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_gemini.AIRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful AI text generation",
                        "schema": {
                            "$ref": "#/definitions/internal_gemini.AIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error during AI generation",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/ask/event/{id}": {
            "post": {
                "description": "Generate text using Gemini AI with event context and chat history",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "Generate AI text response for a specific event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Thread ID for chat history",
                        "name": "thread_id",
                        "in": "query"
                    },
                    {
                        "description": "AI Text Generation Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_gemini.AIRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful AI text generated",
                        "schema": {
                            "$ref": "#/definitions/internal_gemini.AIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error during AI generation",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_holycann_cultour-backend_internal_response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Actual response data"
                },
                "error": {
                    "description": "Detailed error information (only populated for error responses)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.ErrorDetails"
                        }
                    ]
                },
                "message": {
                    "description": "Human-readable message",
                    "type": "string"
                },
                "pagination": {
                    "description": "Pagination information (optional)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.Pagination"
                        }
                    ]
                },
                "request_id": {
                    "description": "Unique request identifier for tracing",
                    "type": "string"
                },
                "status": {
                    "description": "Status of the response (success/error)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.ResponseStatus"
                        }
                    ]
                }
            }
        },
        "github_com_holycann_cultour-backend_internal_response.ErrorDetails": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Machine-readable error code",
                    "type": "string"
                },
                "details": {
                    "description": "Detailed error description",
                    "type": "string"
                }
            }
        },
        "github_com_holycann_cultour-backend_internal_response.Pagination": {
            "type": "object",
            "properties": {
                "has_next_page": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_holycann_cultour-backend_internal_response.ResponseStatus": {
            "type": "string",
            "enum": [
                "success",
                "error"
            ],
            "x-enum-varnames": [
                "StatusSuccess",
                "StatusError"
            ]
        },
...
        "internal_gemini.AIRequest": {
            "type": "object",
            "required": [
                "prompt"
            ],
            "properties": {
                "prompt": {
                    "description": "Prompt is the text input from the user for the AI\nExample: \"Explain quantum computing in simple terms\"\nRequired: true\nMin length: 1\nMax length: 2000",
                    "type": "string",
                    "maxLength": 2000,
                    "minLength": 1,
                    "example": "Explain quantum computing in simple terms"
                }
            }
        },
        "internal_gemini.AIResponse": {
            "type": "object",
            "properties": {
                "metadata": {
                    "description": "Metadata about the generated AI content",
                    "type": "object",
                    "properties": {
                        "length": {
                            "description": "Length of the generated text",
                            "type": "integer",
                            "example": 250
                        },
                        "tokens_used": {
                            "description": "Tokens used in the generation process",
                            "type": "integer",
                            "example": 60
                        }
                    }
                },
                "response": {
                    "description": "Generated text from the AI",
                    "type": "string",
                    "example": "Quantum computing is a type of computing that uses quantum-mechanical phenomena..."
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Insert Your JWT Token. Do NOT include \"Bearer \" prefix. Example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}